code cleanup and reorganization
