+++ packages/browser-sdk/src/toolbar/Features.css
@@ -1,0 +1,69 @@
+.search-input {
+  background: transparent;
+  border: none;
+  border-bottom: 1px solid var(--border-color);
+  color: white;
+  padding: 0 0 6px 0;
+  font-family: monospace;
+  margin-bottom: 4px;
+  width: 100%;
+}
+
+.features-table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.features-table tr td {
+  padding: 6px 6px 4px 0;
+}
+
+.reset {
+  color: var(--brand300);
+}
+
+input[type="search"]::-webkit-search-cancel-button {
+  -webkit-appearance: none;
+  display: inline-block;
+  width: 8px;
+  height: 8px;
+  margin-left: 10px;
+  background: linear-gradient(
+      45deg,
+      rgba(0, 0, 0, 0) 0%,
+      rgba(0, 0, 0, 0) 43%,
+      #fff 45%,
+      #fff 55%,
+      rgba(0, 0, 0, 0) 57%,
+      rgba(0, 0, 0, 0) 100%
+    ),
+    linear-gradient(
+      135deg,
+      transparent 0%,
+      transparent 43%,
+      #fff 45%,
+      #fff 55%,
+      transparent 57%,
+      transparent 100%
+    );
+  cursor: pointer;
+}
+
+.feature-name-cell {
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  width: auto;
+}
+
+.feature-reset-cell {
+  min-width: 38px;
+}
+
+.feature-link {
+  color: var(--text-color);
+  text-decoration: none;
+  &:hover {
+    text-decoration: underline;
+  }
+}
