+++ packages/browser-sdk/src/toolbar/Switch.css
