+++ packages/browser-sdk/src/toolbar/Toolbar.css
@@ -202,7 +202,7 @@ input[type="search"]::-webkit-search-cancel-button {
 }
 
 .feature-reset-cell {
-  min-width: "38px";
+  min-width: 38px;
 }
 
 .feature-link {
