+++ packages/browser-sdk/src/toolbar/index.css
@@ -1,0 +1,3 @@
+@import url(./Toolbar.css);
+@import url(./Features.css);
+@import url(./Switch.css);
